backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt
#
appdirs
    # via pooch
audioread
    # via librosa
certifi
    # via requests
cffi
    # via soundfile
charset-normalizer
    # via requests
click
    # via
    #   click-default-group
    #   cloup
    #   manim
click-default-group
    # via manim
cloup
    # via manim
colorama
    # via rich
colour
    # via manim
commonmark
    # via rich
decorator
    # via
    #   librosa
    #   manim
ffmpeg-python
    # via code-video-generator (setup.py)
future
    # via ffmpeg-python
glcontext
    # via moderngl
idna
    # via requests
joblib
    # via
    #   librosa
    #   scikit-learn
librosa
    # via code-video-generator (setup.py)
llvmlite
    # via numba
#manim==0.10.0
    # via code-video-generator (setup.py)
manimpango
    # via manim
mapbox-earcut
    # via manim
moderngl
    # via
    #   manim
    #   moderngl-window
moderngl-window
    # via manim
multipledispatch
    # via pyrr
networkx
    # via manim
numba
    # via
    #   librosa
    #   resampy
numpy
    # via
    #   librosa
    #   manim
    #   mapbox-earcut
    #   moderngl-window
    #   numba
    #   pyrr
    #   resampy
    #   scikit-learn
    #   scipy
packaging
    # via
    #   librosa
    #   pooch
pillow
    # via
    #   manim
    #   moderngl-window
pooch
    # via librosa
pycairo
    # via manim
pycparser
    # via cffi
pydub
    # via manim
pyglet
    # via
    #   code-video-generator (setup.py)
    #   moderngl-window
pygments
    # via
    #   manim
    #   rich
pyparsing
    # via packaging
pyrr
    # via moderngl-window
#requests==2.26.0
    # via
    #   manim
    #   pooch
resampy
    # via librosa
rich
    # via manim
scikit-learn
    # via librosa
scipy
    # via
    #   librosa
    #   manim
    #   resampy
    #   scikit-learn
screeninfo
    # via manim
six
    # via resampy
soundfile
    # via librosa
threadpoolctl
    # via scikit-learn
tqdm
    # via manim
urllib3
    # via requests
watchdog
    # via manim
wrapt
    # via code-video-generator (setup.py)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
Flask
requests
configparser
manim==0.18.1
manim_voiceover== 0.3.7
setuptools
